Bio inspired computing, Matlab code.
These two projects are assignments completed for an advance computing
course taken at the ANU.

The genetic algorithm uses a memetic algorithm to optimize the weights
of a multi-layer perceptron and a cascor cascade network.

Both architectures are used to predict Diabetes Mellitus within patients
given an input of seven medical  readings.